Frontend | moved the layout of explore pages to a separate file
